% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendPlot.R
\name{dendPlot}
\alias{dendPlot}
\title{Plot Pretty Dendrograms}
\usage{
dendPlot(dend, k = NULL, h = NULL, clusters = NULL,
  labels = clusters, groupLabels = NULL, branches_lwd = c(1, 1, 1),
  branches_lty = c(1, 1, 1), labels_cex = 0.5, do.par = F,
  col = RColorBrewer::brewer.pal(9, "Set1")[-c(6:7)], ...)
}
\arguments{
\item{dend}{object of class 'dendrogram’ or ‘hclust’}

\item{k}{number of groups (passed to ‘cutree’).  Default: NULL}

\item{h}{height at which to cut tree (passed to ‘cutree’). Default: NULL}

\item{clusters}{a character or integer vector of cluster ids. If provided, the branches will be coloured according to these clusters. The length must match the number of leaves in the dendrogram (ie number of features). If the vector is named, the function will check the order of clusters and reorder it according to the leaves in the dendrogram if necessary. If the vector is such that the names are the cluster ids, instead of the feature ids, then the function will first flip the names in the vector and then proceed. If the vector is not named, an already correct order is assumed. Default: NULL}

\item{labels}{rename the labels on the dendrogram. The order of labels is checked as above for <clusters>. If <clusters> (the default), the labels will be replaced with the ID of the cluster in question. Default: clusters}

\item{groupLabels}{if TRUE add numeric group label. See dendextend::color_branches for details. Default: NULL}

\item{branches_lwd}{lwd of branches., Default: c(1, 1, 1)}

\item{branches_lty}{lty of branches. Default: c(1, 1, 1)}

\item{labels_cex}{cex of labels. Default: 0.5}

\item{do.par}{logical indicating whether to use default margin layout for base plot. Default: F}

\item{col}{vector of colours to match clusters, either defined by k, h, or clusters arguments. If the vector is shorter than the number of clusters, colours will be recycled; if it is longer, later colours will be ignored. Default: RColorBrewer::brewer.pal(9, "Set1")[-c(6:7)]}

\item{...}{other arguments passed to base plot (e.g. main = 'My Dendrogram')}
}
\value{
base plot; if output is saved, then return an invisible list containing the dendrogram object and the clusters and labels vectors in case they were reordered/changed in function call.
}
\description{
Plot pretty dendrograms. Change labels, label colours, branch colours, line-widths etc. Change the colours according to k desired number of clusters, or according to h desired height at which to cut the three, or vector <clusters> of the same length as features (leaves) in the tree and with the clusters pre-defined.
}
\seealso{
\code{\link[RColorBrewer]{RColorBrewer}}
 \code{\link[graphics]{par}}
 \code{\link[stats]{setNames}}
 \code{\link[dendextend]{color_branches}},\code{\link[dendextend]{set}},\code{\link[dendextend]{color_labels}},\code{\link[dendextend]{labels_colors}},\code{\link[dendextend]{get_leaves_branches_col}}
 \code{\link[base]{labels}}
}
