% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal_dims.R
\name{equal_cols}
\alias{equal_cols}
\title{Subset matrices to contain the same columns}
\usage{
equal_cols(..., cols = NULL)
}
\arguments{
\item{...}{matrices to subset}

\item{cols}{character vector; additional colnames to subset matrices by. Default: NULL}
}
\value{
matrices with equal columns
}
\description{
subset matrices to contain the same columns
}
\examples{
\dontrun{
 c(m1, m2) \%<-\% equal_cols(m1, m2, cols = NULL)
}
}
