% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal_dims.R
\name{equal_dims}
\alias{equal_dims}
\title{Subset matrices to contain the same columns and rows}
\usage{
equal_dims(..., rows = NULL, cols = NULL)
}
\arguments{
\item{...}{matrices to subset}

\item{rows}{character vector; additional rownames to subset matrices by. Default: NULL}

\item{cols}{character vector; additional colnames to subset matrices by. Default: NULL}
}
\value{
matrices with equal dimensions and the same rows and columns
}
\description{
subset matrices to contain the same columns and rows
}
\examples{
\dontrun{
 c(m1, m2) \%<-\% equal_dims(m1, m2, rows = NULL, cols = sample(colnames(m1), 5))
}
}
