% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal_dims.R
\name{equal_rows}
\alias{equal_rows}
\title{Subset matrices to contain the same rows}
\usage{
equal_rows(..., rows = NULL)
}
\arguments{
\item{...}{matrices to subset}

\item{rows}{character vector; additional rownames to subset matrices by. Default: NULL}
}
\value{
matrices with equal rows
}
\description{
subset matrices to contain the same rows
}
\examples{
\dontrun{
 c(m1, m2) \%<-\% equal_rows(m1, m2, rows = NULL)
}
}
