% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{get_colors}
\alias{get_colors}
\title{Get colors for plotting functions}
\usage{
get_colors(colors, num.colors = 1, reverse = FALSE, gradient = FALSE)
}
\arguments{
\item{colors}{The name of the desired color class or a vector of colors.
See details of \link{jtools_colors}.}

\item{num.colors}{How many colors should be returned? Default is 1.}

\item{reverse}{Should the colors be returned in reverse order, compared to
normal? Default is FALSE.}

\item{gradient}{Return endpoints for a gradient? Default is FALSE. If TRUE,
\code{num.colors} is ignored.}
}
\description{
This is a helper function that provides hex color codes for
\code{jtools}, \code{interactions}, and perhaps other packages.
}
