% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-exts.R
\docType{data}
\name{draw_key_vpath_h}
\alias{draw_key_vpath_h}
\alias{draw_key_pointrange_h}
\alias{geom_pointrange_h}
\alias{GeomPointrangeh}
\alias{geom_linerange_h}
\alias{GeomLinerangeh}
\alias{position_dodgev}
\alias{PositionDodgev}
\title{ggplot2 extensions}
\usage{
draw_key_vpath_h(data, params, size)

draw_key_pointrange_h(data, params, size)

geom_pointrange_h(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., fatten = 4, na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE)

geom_linerange_h(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)

position_dodgev(height = NULL)
}
\arguments{
\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame.}, and
will be used as the layer data.}

\item{params}{Additional parameters to the \code{geom} and \code{stat}.}

\item{size}{size.}

\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{other arguments passed on to layer. These are often aesthetics,
used to set an aesthetic to a fixed value, like color = "red" or size = 3.
They may also be parameters to the paired geom/stat.}

\item{fatten}{A multiplicative factor used to increase the size of the
middle bar in \code{geom_crossbar()} and the middle point in
\code{geom_pointrange()}.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}

\item{height}{Dodging height, when different to the height of the individual
elements. This is useful when you want to align narrow geoms with taller
geoms.}
}
\description{
These are lightly documented tweaks to ggplot2 geoms for
jtools functions.
}
\keyword{datasets}
