% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_summ.R
\name{tidy.summ}
\alias{tidy.summ}
\alias{glance.summ.svyglm}
\alias{glance.summ.lm}
\alias{glance.summ.glm}
\alias{glance.summ.merMod}
\title{Broom extensions for summ objects}
\usage{
tidy.summ(x, ...)

glance.summ.svyglm(x, ...)

glance.summ.lm(x, ...)

glance.summ.glm(x, ...)

glance.summ.merMod(x, ...)
}
\arguments{
\item{x}{The \code{summ} object.}

\item{...}{Other arguments (usually ignored)}
}
\value{
A data.frame with columns matching those appropriate for the model
type per \code{glance} documentation.
}
\description{
These are functions used for compatibility with broom's tidying
functions to facilitate use with huxreg, thereby making
\code{\link{export_summs}} works.
}
\seealso{
\code{\link[broom]{glance}}
}
