% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_apa.R
\name{add_gridlines}
\alias{add_gridlines}
\alias{add_x_gridlines}
\alias{add_y_gridlines}
\alias{drop_gridlines}
\alias{drop_x_gridlines}
\alias{drop_y_gridlines}
\title{Add and remove gridlines}
\usage{
add_gridlines(x = TRUE, y = TRUE, minor = TRUE)

add_x_gridlines(minor = TRUE)

add_y_gridlines(minor = TRUE)

drop_gridlines(x = TRUE, y = TRUE, minor.only = FALSE)

drop_x_gridlines(minor.only = FALSE)

drop_y_gridlines(minor.only = FALSE)
}
\arguments{
\item{x}{Apply changes to the x axis?}

\item{y}{Apply changes to the y axis?}

\item{minor}{Add minor gridlines in addition to major?}

\item{minor.only}{Remove only the minor gridlines?}
}
\description{
These are convenience wrappers for editing \code{\link[ggplot2:theme]{ggplot2::theme()}}'s
\code{panel.grid.major} and \code{panel.grid.minor} parameters with sensible
defaults.
}
