% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions_deps.R
\name{interact_plot}
\alias{interact_plot}
\alias{cat_plot}
\alias{sim_slopes}
\alias{johnson_neyman}
\alias{probe_interaction}
\title{Deprecated interaction functions}
\usage{
interact_plot(...)

cat_plot(...)

sim_slopes(...)

johnson_neyman(...)

probe_interaction(...)
}
\arguments{
\item{...}{arguments are ignored}
}
\description{
These functions are now part of the \pkg{interactions}
package.
}
