% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal_dims.R
\name{intersect_vctrs}
\alias{intersect_vctrs}
\title{Intersection of N atomic vectors}
\usage{
intersect_vctrs(...)
}
\arguments{
\item{...}{atomic vectors to intersect}
}
\value{
intersect of atomic vectors
}
\description{
Return the intersection of any number of atomic vectors provided on input
}
