% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ.R
\name{j_summ}
\alias{j_summ}
\title{Regression summaries with options}
\usage{
j_summ(model, ...)
}
\arguments{
\item{model}{A \code{lm}, \code{glm}, \code{\link[survey]{svyglm}}, or
\code{\link[lme4]{merMod}} object.}

\item{...}{Other arguments to be passed to the model.specific function.}
}
\description{
\code{j_summ} is an alias for \code{summ}.
To get specific documentation, choose the appropriate link to the
type of model that you want to summarize from the details section.
}
\details{
\itemize{

\item \code{\link{summ.lm}}
  \item \code{\link{summ.glm}}
  \item \code{\link{summ.svyglm}}
  \item \code{\link{summ.merMod}}

}
}
