% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ.R
\name{knit_print.summ.lm}
\alias{knit_print.summ.lm}
\alias{knit_print.summ.glm}
\title{knitr methods for summ}
\usage{
knit_print.summ.lm(x, ...)

knit_print.summ.glm(x, ...)
}
\arguments{
\item{x}{The \code{summ} object}

\item{...}{Ignored.}
}
\description{
There's no reason for end users to utilize these functions,
but CRAN requires it to be documented.
}
