% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ.R, R/summ_rq.R
\name{knit_print.summ.lm}
\alias{knit_print.summ.lm}
\alias{knit_print.summ.glm}
\alias{knit_print.summ.svyglm}
\alias{knit_print.summ.merMod}
\alias{knit_print.summ.rq}
\title{knitr methods for summ}
\usage{
\method{knit_print}{summ.lm}(x, options = NULL, ...)

\method{knit_print}{summ.glm}(x, options = NULL, ...)

\method{knit_print}{summ.svyglm}(x, options = NULL, ...)

\method{knit_print}{summ.merMod}(x, options = NULL, ...)

\method{knit_print}{summ.rq}(x, options = NULL, ...)
}
\arguments{
\item{x}{The \code{summ} object}

\item{options}{Chunk options.}

\item{...}{Ignored.}
}
\description{
There's no reason for end users to utilize these functions,
but CRAN requires it to be documented.
}
