% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programming_helpers.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Not \code{\%in\%}}
\usage{
x \%nin\% table
}
\arguments{
\item{x}{An object}

\item{table}{The object you want see if \code{x} is not in}
}
\value{
A logical vector
}
\description{
This function does the very opposite of \code{\%in\%}
}
\seealso{
Other subsetters: 
\code{\link{\%not\%}()}
}
\concept{subsetters}
