% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programming_helpers.R
\name{num_print}
\alias{num_print}
\title{Numbering printing with signed zeroes and trailing zeroes}
\usage{
num_print(x, digits = getOption("jtools-digits", 2), format = "f")
}
\arguments{
\item{x}{The number(s) to print}

\item{digits}{Number of digits past the decimal to print}

\item{format}{equal to \code{"d"} (for integers),
\code{"f"}, \code{"e"}, \code{"E"}, \code{"g"}, \code{"G"}, \code{"fg"} (for reals). Default is \code{"f"}}
}
\description{
This function will print exactly the amount of digits requested
as well as signed zeroes when appropriate (e.g, \code{-0.00}).
}
