% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_slopes.R
\name{plot.sim_slopes}
\alias{plot.sim_slopes}
\title{Plot coefficients from simple slopes analysis}
\usage{
\method{plot}{sim_slopes}(x, ...)
}
\arguments{
\item{x}{A \code{\link[jtools:sim_slopes]{jtools::sim_slopes()}} object.}

\item{...}{arguments passed to \code{\link[jtools:plot_coefs]{jtools::plot_coefs()}}}
}
\description{
This creates a coefficient plot to visually summarize the
results of simple slopes analysis.
}
