% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_summ.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tidy}
\alias{glance}
\alias{as_tibble}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{tidy}}, \code{\link[generics]{glance}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}
}}

