% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.R
\name{reorder}
\alias{reorder}
\title{Reorder matrices on the fly, FLexiblY.}
\usage{
reorder(m = NULL, cr = NULL, col = T, row = T,
  hc.method = "average", cor.method = "pearson", compute.dist = T,
  dist.method = "euclidean")
}
\arguments{
\item{m}{matrix to be reordered Default: NULL}

\item{cr}{correlation matrix to be reordered. This will allow (1) the correlation matrix to be used, rather than computed, in defining the hierarchical clustering order and (2) if row and col are both set to TRUE, the clustering will only be run once and the ordering applied to both rows and columns. Default: NULL}

\item{col}{logical indicating whether to reorder columns. Default: TRUE}

\item{row}{logical indicating whether to reorder rows. Default: TRUE}

\item{hc.method}{a character string indicating which agglomeration method to use. Default: 'average'}

\item{cor.method}{a character string indicating which correlation coefficient is to be computed. Default: 'pearson'}

\item{compute.dist}{a boolean value indicating whether a distance measure should be computed from the correlation metric. If FALSE, distances are computed from the correlation matrix directly. Default: T}

\item{dist.method}{a character string specifying the distance metric to be used for hierarchical clustering. Default: ''euclidean'}
}
\value{
an ordered matrix with the same number of rows and columns as there were columns in the original matrix.
}
\description{
Calls the jtools function HC() to order the columns and rows by hierarchical clustering, again with parameters specified by the user (defaults are recommended for single-cell expression (RNAseq) data).
}
