% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_slopes.R
\name{sim_slopes}
\alias{sim_slopes}
\title{Perform a simple slopes analysis.}
\usage{
sim_slopes(formula, data, pred, modx, modxvals = NULL, centered = NULL,
  weights = NULL, robust = FALSE, robust.type = "HC3", summary = FALSE)
}
\arguments{
\item{formula}{A regression formula, like would be provided to \code{lm()}, in
quotation marks. This should not include the interaction term. Alternately, give
an \code{lm} object and the formula will be extracted. Interaction terms will be ignored.}

\item{data}{A data frame.}

\item{pred}{The predictor variable involved in the interaction in quotes.}

\item{modx}{The moderator variable involved in the interaction in quotes.}

\item{modxvals}{For which two values of the moderator should simple slopes analysis
be performed? Default is \code{NULL}. If \code{NULL}, then the customary +/-
1 standard deviation from the mean values are used. This would not be appropriate
in the case of a binary moderator, however, as well as in some other use cases.}

\item{centered}{A vector of quoted variable names that are to be mean-centered.}

\item{weights}{If weights are being used, provide the variable name where they are stored.}

\item{robust}{Logical. If \code{TRUE}, computes heteroskedasticity-robust standard errors.}

\item{robust.type}{Type of heteroskedasticity-robust standard errors to use if \code{robust=TRUE}.
See details of \code{\link{j_summ}} for more on options.}

\item{summary}{If \code{TRUE}, prints a model summary using \code{\link{j_summ}}
for the non-mean-adjusted model.}
}
\description{
\code{sim_slopes()} conducts a simple slopes analysis for the purposes of understanding
interaction effects in linear regression.
}
\examples{
# Using a fitted model as formula input
fit <- lm(Income ~ `HS Grad` + Murder + Illiteracy,
  data=as.data.frame(state.x77))
sim_slopes(formula=fit, pred="Murder", modx="Illiteracy",
  centered=c("Murder","Illiteracy"), summary=TRUE)

}

