% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programming_helpers.R
\name{\%not\%}
\alias{\%not\%}
\alias{\%just\%}
\alias{\%not\%.default}
\alias{\%not\%.data.frame}
\alias{\%not\%.matrix}
\alias{\%just\%.default}
\alias{\%just\%.data.frame}
\alias{\%just\%.matrix}
\title{Subsetting operators}
\usage{
x \%not\% y

x \%just\% y

\method{\%not\%}{default}(x, y)

\method{\%not\%}{data.frame}(x, y)

\method{\%not\%}{matrix}(x, y)

\method{\%just\%}{default}(x, y)

\method{\%just\%}{data.frame}(x, y)

\method{\%just\%}{matrix}(x, y)
}
\arguments{
\item{x}{Object to subset}

\item{y}{List of items to include if they are/aren't in \code{x}}
}
\value{
All of \code{x} that are in \code{y} (\code{\%just\%}) or all of \code{x} that are not in
\code{y} (\code{\%not\%}).
}
\description{
\code{\%just\%} and \code{\%not\%} are subsetting convenience functions
for situations when you would do \code{x[x \%in\% y]} or \code{x[x \%nin\% y]}. See
details for behavior when \code{x} is a data frame or matrix.
}
\details{
The behavior of %not% and %just% are different when you're subsetting
data frames or matrices. The subset \code{y} in this case is interpreted as
column names or indices.
}
\examples{

 x <- 1:5
 y <- 3:8
 
 x \%just\% y # 3 4 5
 x \%not\% y # 1 2
 
 mtcars \%just\% c("mpg", "qsec", "cyl") # keeps only columns with those names
 mtcars \%not\% 1:5 # drops columns 1 through 5
 
 
}
