% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svycor.R
\name{svycor}
\alias{svycor}
\title{Calculate Pearson correlations with complex survey data}
\usage{
svycor(formula, design, na.rm = FALSE, sig.stats = FALSE, bootn = 1000,
  mean1 = TRUE, ...)
}
\arguments{
\item{formula}{A formula specifying the terms to correlate.}

\item{design}{The \code{survey.design} or \code{svyrep.design} object.}

\item{na.rm}{Logical. Should cases with missing values be dropped?}

\item{...}{Additional arguments passed to \code{\link[survey]{svyvar}}.}
}
\description{
\code{svycor} extends the \code{survey} package by calculating correlations
with syntax similar to the original package, which for reasons unknown lacks
such a function.
}
\details{

}

