% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programming_helpers.R
\name{wrap_str}
\alias{wrap_str}
\alias{cat_wrap}
\alias{warn_wrap}
\alias{stop_wrap}
\alias{msg_wrap}
\title{\code{cat}, \code{message}, \code{warning}, and \code{stop} wrapped to fit the console's
width.}
\usage{
wrap_str(..., sep = "")

cat_wrap(..., brk = "")

warn_wrap(..., brk = "\\n", call. = FALSE)

stop_wrap(..., brk = "\\n", call. = FALSE)

msg_wrap(..., brk = "\\n")
}
\arguments{
\item{...}{Objects to print}

\item{sep}{Separator between \code{...}, Default: ''}

\item{brk}{What should the last character of the message/warning/error be?
Default is \code{"\n"}, meaning the console output ends with a new line.}

\item{call.}{Here for legacy reasons. It is ignored.}
}
\description{
These are convenience functions that format printed output to
fit the width of the user's console.
}
