% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{wtd.sd}
\alias{wtd.sd}
\title{Weighted standard deviation calculation}
\usage{
wtd.sd(x, weights)
}
\arguments{
\item{x}{A vector of values for which you want the standard deviation}

\item{weights}{A vector of weights equal in length to \code{x}}
}
\description{
This function calculates standard deviations with weights and
is a counterpart to the built-in \code{weighted.mean} function.
}
