% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_summ.R, R/simple_slopes.R, R/summ_rq.R
\name{tidy.summ}
\alias{tidy.summ}
\alias{glance.summ.lm}
\alias{glance.summ.glm}
\alias{glance.summ.svyglm}
\alias{glance.summ.merMod}
\alias{tidy.sim_slopes}
\alias{glance.summ.rq}
\title{Broom extensions for summ objects}
\usage{
tidy.summ(x, conf.int = FALSE, conf.level = 0.95, ...)

glance.summ.lm(x, ...)

glance.summ.glm(x, ...)

glance.summ.svyglm(x, ...)

glance.summ.merMod(x, ...)

tidy.sim_slopes(x, conf.int = FALSE, conf.level = 0.95, ...)

glance.summ.rq(x, ...)
}
\arguments{
\item{x}{The \code{summ} object.}

\item{conf.int}{Include confidence intervals? Default is FALSE.}

\item{conf.level}{How wide confidence intervals should be, if requested.
Default is .95.}

\item{...}{Other arguments (usually ignored)}
}
\value{
A data.frame with columns matching those appropriate for the model
type per \code{glance} documentation.
}
\description{
These are functions used for compatibility with broom's tidying
functions to facilitate use with huxreg, thereby making
\code{\link{export_summs}} works.
}
\seealso{
\code{\link[broom]{glance}}
}
