% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact_plot.R
\name{interact_plot}
\alias{interact_plot}
\title{Plot interaction effects in regression models}
\usage{
interact_plot(model, pred, modx, modxvals = NULL, interval = FALSE,
  int.type = c("confidence", "prediction"), centered = NULL,
  int.width = 0.975, x.label = NULL, y.label = NULL, mod.labels = NULL,
  main.title = NULL, legend.main = NULL, color.class = "Set2")
}
\arguments{
\item{model}{A regression model of type \code{lm}, \code{glm}, or
\code{\link[survey]{svyglm}}. It should contain the interaction of interest.}

\item{pred}{The name of the predictor variable involved in the interaction in
quotations.}

\item{modx}{The name of the moderator variable involved in the interaction in
quotations.}

\item{modxvals}{For which values of the moderator should lines be plotted?
Default is \code{NULL}. If \code{NULL}, then the customary +/- 1 standard
deviation from the mean values are used for continuous moderators. If the
moderator is a factor variable and \code{modxvals} is \code{NULL}, each level
of the factor is included.}

\item{interval}{Logical. If \code{TRUE}, plots confidence/prediction intervals
the line using \code{\link[ggplot2]{geom_ribbon}}.}

\item{int.type}{Type of interval to plot. Options are "confidence" or "prediction".
Default is confidence interval.}

\item{centered}{A vector of quoted variable names that are to be mean-centered.
If \code{NULL}, all non-focal variables are mean-centered. If variables are specified,
then no others are centered. Alternately, "all" will center focal variables
as well.}

\item{int.width}{How large should the interval be, relative to the standard error?
The default, .975, corresponds to roughly 1.96 standard errors and a .05 alpha
level for values outside the range.}

\item{x.label}{A character object specifying the desired x-axis label. If \code{NULL},
the variable name is used.}

\item{y.label}{A character object specifying the desired x-axis label. If \code{NULL},
the variable name is used.}

\item{mod.labels}{A character vector of labels for each level of the moderator values,
provided in the same order as the \code{modxvals} argument. If \code{NULL},
the valeus themselves are used as labels unless \code{modxvals} is also \code{NULL}.
In that case, "+1 SD" and "-1 SD" are used.}

\item{main.title}{A character object that will be used as an overall title for the
plot. If \code{NULL}, no main title is used.}

\item{legend.main}{A character object that will be used as the title that appears
above the legend. If \code{NULL}, the name of the moderating variable is used.}

\item{color.class}{No options available for this in-development feature.}
}
\value{
The functions returns a \code{ggplot} object, which can be treated like
  a user-created plot and expanded upon as such.
}
\description{
\code{interact_plot()} plots regression lines at user-specified levels of a moderator
variable to explore interactions. The plotting is done with \code{ggplot2} rather than
base graphics, which some similar functions use.
}
\details{
This function provides a means for plotting conditional effects for the
  purpose of exploring interactions in the context of regression. You must have the
  package \code{ggplot2} installed to benefit from these plotting functions.

  The function is designed for two-way interactions. For additional terms, the
  \code{\link[effects]{effects}} package may be better suited to the task.

  This function has not been extensively tested with non-linear models and may
  display unusual behavior and/or errors when used with them.
}
\examples{
# Using a fitted lm model
states <- as.data.frame(state.x77)
states$HSGrad <- states$`HS Grad`
fit <- lm(Income ~ HSGrad + Murder*Illiteracy,
  data = states)
interact_plot(model = fit, pred = "Murder",
  modx = "Illiteracy")

# Using interval feature
fit <- lm(accel ~ mag*dist, data=attenu)
interact_plot(fit, pred = "mag", modx = "dist", interval = TRUE,
  int.type = "confidence", int.width = .8)


}
\references{
Bauer, D. J., & Curran, P. J. (2005). Probing interactions in fixed and multilevel
 regression: Inferential and graphical techniques. \emph{Multivariate Behavioral
 Research}, \emph{40}(3), 373-400.

Cohen, J., Cohen, P., West, S. G., & Aiken, L. S. (2003). \emph{Applied multiple
regression/correlation analyses for the behavioral sciences} (3rd ed.).
Mahwah, NJ: Lawerence Erlbaum Associates, Inc.
}
\seealso{
\code{\link[rockchalk]{plotSlopes}} performs a similar function, but
  but with the base graphics package--this function is meant, in part, to simulate
  its features.

  \code{\link{sim_slopes}} performs a simple slopes analysis with a similar
  argument syntax to this function.
}
\author{
Jacob Long <\email{long.1377@osu.edu}>
}

