% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_summary.R
\name{j_summ}
\alias{j_summ}
\title{Regression summaries with options}
\usage{
j_summ(lm, stdbeta = FALSE, vifs = FALSE, robust = FALSE,
  robust.type = "HC3", digits = 5)
}
\arguments{
\item{lm}{A \code{lm}, \code{glm}, or \code{\link[survey]{svyglm}} object.}

\item{stdbeta}{If \code{TRUE}, adds a column to output with standardized regression
coefficients. Default is \code{FALSE}.}

\item{vifs}{If \code{TRUE}, adds a column to output with variance inflation factors
(VIF). Default is \code{FALSE}.}

\item{robust}{If \code{TRUE}, reports heteroskedasticity-robust standard errors
instead of conventional SEs. These are also known as Huber-White standard errors.
Default is \code{FALSE}.
This requires the \code{sandwich} and \code{lmtest} packages to compute the
 standard errors.}

\item{robust.type}{Only used if \code{robust=TRUE}. Specificies the type of
robust standard errors to be used by \code{sandwich}. By default, set to \code{"HC3"}
. See details for more on options.}

\item{digits}{An integer specifying the number of digits past the decimal to report in
the output. Default is 5.}
}
\description{
\code{j_summ()} prints output for a regression model in a fashion similar to
\code{summary()}, but formatted differently with more options.
}
\details{
Regardless of options selected, this function will print the following items to the console:
\itemize{
  \item The sample size
  \item The name of the outcome variable
  \item The number of predictors used
  \item The (Pseudo-)R-squared value (plus adjusted R-squared if OLS regression).
  \item A table with regression coefficients, standard errors, t-values, and p-values.
}

 There are several options available for \code{robust.type}. The heavy lifting is done by
 \code{\link[sandwich]{vcovHC}}, where those are better described. Put simply, you may choose
 from \code{"HC0"} to \code{"HC5"}. Based on the recommendation of the developers of
 \code{sandwich}, the default is set to \code{"HC3"}. Stata's default is \code{"HC1"}, so
 that choice may be better if the goal is to replicate Stata's output. Any option that
 is understood by \code{vcovHC} will be accepted.
}
\examples{
# Create lm object
fit <- lm(Income ~ `HS Grad` + Illiteracy + Murder, data=as.data.frame(state.x77))

# Print the output with standardized coefficients and 2 digits past the decimal
j_summ(fit, stdbeta=TRUE, digits=2)

}
\author{
Jacob Long <\email{long.1377@osu.edu}>
}

