% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j_summary.R
\name{j_summ}
\alias{j_summ}
\title{Regression summaries with options}
\usage{
j_summ(lm, stdbeta = FALSE, vifs = FALSE, robust = FALSE,
  robust.type = "HC3", digits = 3, model.info = TRUE, model.fit = TRUE,
  model.check = FALSE)
}
\arguments{
\item{lm}{A \code{lm}, \code{glm}, or \code{\link[survey]{svyglm}} object.}

\item{stdbeta}{If \code{TRUE}, adds a column to output with standardized regression
coefficients. Default is \code{FALSE}.}

\item{vifs}{If \code{TRUE}, adds a column to output with variance inflation factors
(VIF). Default is \code{FALSE}.}

\item{robust}{If \code{TRUE}, reports heteroskedasticity-robust standard errors
instead of conventional SEs. These are also known as Huber-White standard errors.
Default is \code{FALSE}.
This requires the \code{sandwich} and \code{lmtest} packages to compute the
 standard errors.}

\item{robust.type}{Only used if \code{robust=TRUE}. Specificies the type of
robust standard errors to be used by \code{sandwich}. By default, set to \code{"HC3"}
. See details for more on options.}

\item{digits}{An integer specifying the number of digits past the decimal to report in
the output. Default is 5.}

\item{model.info}{Toggles printing of basic information on sample size, name of
DV, and number of predictors.}

\item{model.fit}{Toggles printing of R-squared, Adjusted R-squared, Pseudo-R-squared,
and AIC (when applicable).}

\item{model.check}{Toggles whether to perform Breusch-Pagan test for heteroskedasticity
and print number of high-leverage observations. See details for more info.}
}
\description{
\code{j_summ()} prints output for a regression model in a fashion similar to
\code{summary()}, but formatted differently with more options.
}
\details{
By default, this function will print the following items to the console:
\itemize{
  \item The sample size
  \item The name of the outcome variable
  \item The number of predictors used
  \item The (Pseudo-)R-squared value (plus adjusted R-squared if OLS regression).
  \item A table with regression coefficients, standard errors, t-values, and p-values.
}

 There are several options available for \code{robust.type}. The heavy lifting is done by
 \code{\link[sandwich]{vcovHC}}, where those are better described. Put simply, you may choose
 from \code{"HC0"} to \code{"HC5"}. Based on the recommendation of the developers of
 \code{sandwich}, the default is set to \code{"HC3"}. Stata's default is \code{"HC1"}, so
 that choice may be better if the goal is to replicate Stata's output. Any option that
 is understood by \code{vcovHC} will be accepted.

 There are two pieces of information given for \code{model.check}, provided that
 the model is an \code{lm} object. First, a Breusch-Pagan test is performed with
 \code{\link[car]{ncvTest}}, which requires the \code{car} package. This is a
 hypothesis test for which the alternative hypothesis is heteroskedastic errors.
 The test becomes much more likely to be statistically significant as the sample
 size increases; however, the homoskedasticity assumption becomes less important
 to inference as sample size increases (Lumley, Dieher, Emerson, & Lu, 2002).
 Take the result of the test as a cue to check graphical checks rather than a
 definitive decision. Note that the use of robust standard errors can account
  for heteroskedasticity, though some oppose this approach (see King & Roberts, 2015).

 The second piece of information provided by setting \code{model.check} to \code{TRUE}
 is the number of high leverage observations. There are no hard and fast rules
 for determining high leverage either, but in this case it is based on Cook's
 Distance. All Cook's Distance values greater than (4/N) are included in the
 count. Again, this is not a recommendation to locate and remove such observations,
 but rather to look more closely with graphical and other methods.
}
\examples{
# Create lm object
fit <- lm(Income ~ `HS Grad` + Illiteracy + Murder, data=as.data.frame(state.x77))

# Print the output with standardized coefficients and 2 digits past the decimal
j_summ(fit, stdbeta=TRUE, digits=2)


}
\references{
King, G., & Roberts, M. E. (2015). How robust standard errors expose methodological
 problems they do not fix, and what to do about it. \emph{Political Analysis},
  \emph{23}(2), 159–179. \url{https://doi.org/10.1093/pan/mpu015}

Lumley, T., Diehr, P., Emerson, S., & Chen, L. (2002). The Importance of the
Normality Assumption in Large Public Health Data Sets. \emph{Annual Review of
 Public Health}, \emph{23}, 151–169.
 \url{https://doi.org/10.1146/annurev.publhealth.23.100901.140546}
}
\author{
Jacob Long <\email{long.1377@osu.edu}>
}

