% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_slopes.R
\name{sim_slopes}
\alias{sim_slopes}
\title{Perform a simple slopes analysis}
\usage{
sim_slopes(model, pred, modx, modxvals = NULL, centered = NULL,
  robust = FALSE, robust.type = "HC3", cond.int = FALSE, digits = 3)
}
\arguments{
\item{model}{A regression model of type \code{lm} or \code{\link[survey]{svyglm}}.
It should contain the interaction of interest.}

\item{pred}{The predictor variable involved in the interaction.}

\item{modx}{The moderator variable involved in the interaction.}

\item{modxvals}{For which values of the moderator should simple slopes analysis
be performed? Default is \code{NULL}. If \code{NULL}, then the values will be
the customary +/- 1 standard deviation from the mean as well as the mean itself.
There is no specific limit on the number of variables provided. Factor variables
are not particularly suited to simple slopes analysis, but you could have a
numeric moderator with values of 0 and 1 and give \code{c(0,1)} to compare the
slopes at the different conditions.}

\item{centered}{A vector of quoted variable names that are to be mean-centered. If
\code{NULL}, all non-focal predictors are centered. If not \code{NULL}, only
the user-specified predictors are centered. User can also use "none" or "all"
arguments.}

\item{robust}{Logical. If \code{TRUE}, computes heteroskedasticity-robust standard errors.}

\item{robust.type}{Type of heteroskedasticity-robust standard errors to use if \code{robust=TRUE}.
See details of \code{\link{j_summ}} for more on options.}

\item{cond.int}{Should conditional intercepts be printed in addition to the slopes?
Default is \code{FALSE}.}

\item{digits}{How many significant digits after the decimal point should the output
contain?}
}
\value{
\item{slopes}{A table of coefficients for the focal predictor at each value of
 the moderator}
 \item{ints}{A table of coefficents for the intercept at each value of the moderator}
 \item{modxvals}{The values of the moderator used in the analysis}
 \item{mods}{A list containing each regression model created to estimate the conditional
 coefficients.}
}
\description{
\code{sim_slopes()} conducts a simple slopes analysis for the purposes of understanding
interaction effects in linear regression.
}
\details{
This allows the user to perform a simple slopes analysis for the purpose
  of probing interaction effects in a linear regression. Only two-way interactions
  are supported.

  The function accepts a \code{lm} object and uses it to recompute models with
  the moderating variable set to the levels requested. \code{\link[survey]{svyglm}}
   objects are also accepted, though users should be cautioned against using
  simple slopes analysis with non-linear models.
}
\examples{
# Using a fitted model as formula input
fiti <- lm(Income ~ `HS Grad` + Murder*Illiteracy,
  data=as.data.frame(state.x77))
sim_slopes(model=fiti, pred=Murder, modx=Illiteracy)


}
\references{
Bauer, D. J., & Curran, P. J. (2005). Probing interactions in fixed and multilevel
 regression: Inferential and graphical techniques. \emph{Multivariate Behavioral
 Research}, \emph{40}(3), 373-400.

Cohen, J., Cohen, P., West, S. G., & Aiken, L. S. (2003). \emph{Applied multiple
regression/correlation analyses for the behavioral sciences} (3rd ed.).
Mahwah, NJ: Lawerence Erlbaum Associates, Inc.
}
\seealso{
\code{\link{interact_plot}} accepts similar syntax and will plot the
  results with \code{\link[ggplot2]{ggplot}}.

  \code{\link[rockchalk]{testSlopes}} performs a hypothesis test of differences
  and provides Johnson-Neyman intervals.

  \code{\link[pequod]{simpleSlope}} performs a similar analysis and can analyze
  a second moderator.
}
\author{
Jacob Long <\email{long.1377@osu.edu}>
}
